Function numcodecs_reinterpret::reinterpret_array_into

pub fn reinterpret_array_into<'a, T: Copy, U: ArrayDType, D: Dimension>(
    encoded: ArrayView<'_, T, D>,
    reinterpret: impl Fn(T) -> U,
    decoded: AnyArrayViewMut<'a>,
) -> Result<(), ReinterpretCodecError>
where
    U::RawData<ViewRepr<&'a mut ()>>: DataMut,

Reinterpret the data elements of the encoded array using the provided reinterpret closure into the decoded array.

Errors

Errors with

ReinterpretCodecError::MismatchedDecodeIntoDtype if decoded does not contain an array with elements of type U

ReinterpretCodecError::MismatchedDecodeIntoShape if decoded’s shape does not match encoded’s shape